refactor report job search route
